refactor: extraction KioskScoreEntry de CompetitionView (phase 2)
CompetitionView 1800 → 1689 lignes : mode kiosk de saisie des scores extrait dans competition/KioskScoreEntry.tsx (React.memo), styles déplacés. Comportement identique — type-check, build webpack et tests inchangés.
